Revert to working version
Unsure if modularization works, will need to read into documentation. For now, this version has a working program where the game can be played. There are also classes that have the modularization, but putting them together doesnt seem to work. There, finally, is a class that is the previous version that failed to implement the power ups properly, but that is saved as a 'snapshot' to remember the logic where it was for when it is started up again.