Add new mini-app
Create new directory to hold future implementations for a todo-list web application. Current version utilizes local storage for persistence; future versions will utilize a flask-based back-end if possible. That is the main point of this directory--test github pages compatibility with a flask backend for managing databases and collecting/storing information